+++ core/__init__.py
@@ -3,10 +3,9 @@
 """
 from .session import CardSession, ToolCall, SessionManager
 from .render import render_card
-from .normalizer import StreamingTextNormalizer, normalize_stream_text
+from .normalizer import StreamingTextNormalizer
 from .patch import StreamingPatch
 from .lark_card import (
-    build_streaming_card,
     capture_lark_ids_from_response,
     chunk_to_text,
     mutate_card_request_object,
@@ -32,9 +31,7 @@ __all__ = [
     "SessionManager",
     "render_card",
     "StreamingTextNormalizer",
-    "normalize_stream_text",
     "StreamingPatch",
-    "build_streaming_card",
     "capture_lark_ids_from_response",
     "chunk_to_text",
     "mutate_card_request_object",